update ARIA mappings for attributes
While the `meter` role will be defined in ARIA 1.2, the attribute mappings for `min` and `max` are availalbe now, so we should point to those.

This commit also updates the HTML attr links for these attributes.

Additionally, ARIA 1.2 notes that an attribute for `low`, `high` and `optimum` will not be made until ARIA 1.3, if at all.  Those HTML attribute ARIA mappings have also been updated as necessary.