comment out rb and rtc
closes #115

While this PR fills in mappings for `rb` and the `rtc` elements based on how they are presently mapped, they are also being commented out as they are marked as obsolete in the HTML specification.

If consensus can be reached on adding these elements back into the HTML specification, then this PR should be reversed.