Editorial: Clarify reversed attribute
Some of the text for the different platforms didn't align with what occurs in all browsers today (possibly there were variants in the past?)  

All mapping rows are now the same. `reversed` changes the list marker ordering (numeric or alphabetic) to be in reverse order, rather than starting with the lowest (1, A, etc) ordering marker.