add missing label elm to 'button' naming steps
adds label element to `button`, `input type=button|submit|reset|image`.

This is a long overdue update and there could be more clarity here.  However, getting this perfect has long stood in the way of getting this done, and I would rather get this gap resolved to match reality, and then better describe this in a follow revision to these naming steps.