Adds language to expand 'not mapped' definition
closes #151

While there are many elements that would have originally fallen under this issue... on a long enough timeline everything maps to `generic`.

So is the resolution to this issue.  There are a few remaining HTML elements that are not mapped because they are not rendered by browsers.  For the few instances of these where users can force render them (e.g., making a `style` element `display: block` for reasons) then these elements would and should be largely exposed as `generic` elements.  From quick review, the only one in question here of not mapping to `generic` is `link`. But also, if you're force rendering a `link` element... then...