update reference to placeholder in naming computation
related to https://github.com/w3c/accname/issues/17

In terms of how naming is calculated, because `placeholder` is included as the last-resort option for naming, that means that `aria-placeholder` is too.  So, without calling that out explicitly (because we absolutely don't want to make it seem that naming an element via `aria-placeholder` is at all a proper thing to do), instead slightly reword the step to allude to using the placeholder value instead of the attribute itself.  The placeholder text now links to the attribute, which indicates it maps to the aria-placeholder property... so, a very round about way of calling this out, but roundabout on purpose.

The alternative to this is we just list `placeholder` and `aria-placeholder` and indicate that's the order of preference - HTML feature over ARIA if both are present.