Clarification: label element and for attribute
This PR clarifies some of the wording in the `label` element mappings where, for instance, UIA was written in a way that was specifically calling out labelling the `input` element - but there are a number of labelable elements that are not `input`.  Rather than say all of them or even say 'form controls' which is also not necessarily accurate, instead mention and link to the HTML spec's definition and listing of labelable elements.

The `label[for]` attribute was updated to consistently indicate that it's used for accname by all the platforms.  UIA's 'not mapped' text was replaced with a version of the text that existed in the `label` element's UIA mapping.  MSAA/IA2 mappings were updated to reference the labelable element section from HTML.

note: this work was done as I noticed these things needed clarifying while making other additions to #402 today.