Clarify mappings for contenteditable
Per discussion around the minimum role concept, and the WG deciding that contentedtiable should remain unmapped to any specific ARIA role when used on a generic.    per that decision #324 is closed, but the other clarifications that were to be a part of that PR are now in this one.

Additionally, there are other necessary updates needed for these mappings, as the previous ones did not get into the nuance of how to handle nested `contenteditable` attributes.