add computed role mappings for html elements
resolves #464

The initial PR adds the computed role cell to all HTML elements.  Where possible, the cells have been filled in with 'use aria mappings', 'not mapped' or 'see comments' or '???'.  The ??? instances are those where there are host-language roles that need to be determined for elements which are important to accessibility but have no explicit ARIA role to neatly map to.